Pinecone Serverless Reranking/requirements.txt
pinecone>=6.0.1
pinecone-notebooks
pandas
torch
transformers
requests
python-dotenv
